configure paginated queries processing, add support for Pageable.unpaged()

+++ src/main/java/org/springframework/data/aerospike/config/AbstractAerospikeDataConfiguration.java
@@ -57,7 +57,7 @@ public abstract class AbstractAerospikeDataConfiguration extends AerospikeDataCo
         boolean scansEnabled = aerospikeDataSettings().isScansEnabled();
         log.debug("AerospikeDataSettings.scansEnabled: {}", scansEnabled);
         queryEngine.setScansEnabled(scansEnabled);
-        int queryMaxRecords = aerospikeDataSettings().getQueryMaxRecords();
+        long queryMaxRecords = aerospikeDataSettings().getQueryMaxRecords();
         log.debug("AerospikeDataSettings.queryMaxRecords: {}", queryMaxRecords);
         queryEngine.setQueryMaxRecords(queryMaxRecords);
         return queryEngine;
